requirements.txt
python-dotenv==0.17.1
bs4==0.0.1
requests==2.23.0
selenium==4.0.0a1